requirements.txt
APScheduler==3.11.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
gpg==1.24.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
pycodestyle==2.12.1
python-dotenv==1.1.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.25.1
tzlocal==5.3.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel @ file:///opt/homebrew/Cellar/python%403.13/3.13.3/libexec/wheel-0.45.1-py3-none-any.whl#sha256=c00982b077e92504359d52a0482365099ab8c23f26817848cf77c784f6af247c
